fish-id/backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
pydantic
Pillow
torch ; platform_system != 'Windows'  # optional, skip if issues
transformers
rich
python-dotenv
